# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Boxty API
  version: 1.0.0
  description: |
    Boxty — programmable sandbox platform. Run code in cloud with gVisor isolation,
    secrets, cron scheduling, image building, persistent volumes, and web endpoint API.
servers:
  - url: https://api.boxty.dev
    description: Production
  - url: http://localhost:3000
    description: Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from login/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key format "Bearer bxty_..."
  schemas:
...
    Sandbox:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [running, stopped, snapshotted] }
        url: { type: string, format: uri }
        image: { type: string }
        cpu: { type: integer }
        memory: { type: integer }
        gpu: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        timeout: { type: integer }
...
    Secret:
      type: object
      properties:
        name: { type: string }
        created_at: { type: string, format: date-time }
...
    Volume:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        size_gb: { type: integer }
        status: { type: string }
        mount_url: { type: string }
        created_at: { type: string, format: date-time }
    Schedule:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        schedule_type: { type: string, enum: [cron, period] }
        schedule_value: { type: string }
        next_run: { type: string, format: date-time }
        status: { type: string, enum: [active, paused] }
...
paths:
...
  /api/sandboxes:
    post:
      summary: Create sandbox
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image]
              properties:
                image: { type: string, example: "pytorch:latest" }
                cpu: { type: integer, default: 2 }
                memory: { type: integer, default: 4096 }
                gpu: { type: string, nullable: true }
                timeout: { type: integer, default: 3600 }
                disk_size_gb: { type: integer, default: 10 }
                volume: { type: string, nullable: true }
                volume_mount_path: { type: string, default: "/mnt/volume" }
      responses:
        '201':
          description: Sandbox created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
    get:
      summary: List sandboxes
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        '200':
          description: List of sandboxes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Sandbox' }
  /api/sandboxes/{id}:
    get:
      summary: Get sandbox details
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sandbox details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sandbox' }
    delete:
      summary: Stop and delete sandbox
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sandbox stopped
...
  /api/sandboxes/{id}/snapshot:
    post:
      summary: Create snapshot
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        '200':
          description: Snapshot created
  /api/sandboxes/restore:
    post:
      summary: Restore from snapshot
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snapshotKey, image]
              properties:
                snapshotKey: { type: string }
                image: { type: string }
                cpu: { type: integer }
                memory: { type: integer }
                gpu: { type: string }
                secrets: { type: array }
      responses:
        '201':
          description: Sandbox restored
  /api/sandboxes/{id}/forward:
    post:
      summary: Port forwarding
      tags: [Sandboxes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [port]
              properties:
                port: { type: integer }
      responses:
        '200':
          description: Forward URL
...
  /api/workers/register:
    post:
      summary: Register worker
      tags: [Workers]
      security: [{ ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                host: { type: string }
                region: { type: string }
                provider: { type: string }
      responses:
        '201':
          description: Worker registered
  /api/workers:
    get:
      summary: List workers
      tags: [Workers]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Workers list
  /api/workers/{id}/heartbeat:
    post:
      summary: Worker heartbeat
      tags: [Workers]
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cpu: { type: object }
                memory: { type: object }
                sandboxes_running: { type: integer }
      responses:
        '200':
          description: Heartbeat received
...
  /api/secrets:
    post:
      summary: Create secret
      tags: [Secrets]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, value]
              properties:
                name: { type: string }
                value: { type: string }
      responses:
        '201':
          description: Secret created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Secret' }
    get:
      summary: List secrets
      tags: [Secrets]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        '200':
          description: Secrets list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Secret' }
  /api/secrets/{name}:
    delete:
      summary: Delete secret
      tags: [Secrets]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Secret deleted
  /api/volumes:
    post:
      summary: Create volume
      tags: [Volumes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                size_gb: { type: integer, default: 50 }
      responses:
        '201':
          description: Volume created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Volume' }
    get:
      summary: List volumes
      tags: [Volumes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        '200':
          description: Volumes list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Volume' }
  /api/volumes/{id}:
    get:
      summary: Get volume details
      tags: [Volumes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Volume details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Volume' }
    delete:
      summary: Delete volume
      tags: [Volumes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Volume deleted
  /api/volumes/{id}/mount:
    post:
      summary: Mount volume
      tags: [Volumes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sandbox_id, mount_path]
              properties:
                sandbox_id: { type: string }
                mount_path: { type: string }
      responses:
        '200':
          description: Volume mounted
  /api/volumes/{id}/unmount:
    post:
      summary: Unmount volume
      tags: [Volumes]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sandbox_id]
              properties:
                sandbox_id: { type: string }
      responses:
        '200':
          description: Volume unmounted
  /api/schedules:
    post:
      summary: Create schedule
      tags: [Schedules]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, schedule_type, schedule_value, function_name, image, cpu, memory]
              properties:
                name: { type: string }
                schedule_type: { type: string, enum: [cron, period] }
                schedule_value: { type: string }
                function_name: { type: string }
                args: { type: object }
                image: { type: string }
                cpu: { type: integer }
                memory: { type: integer }
                gpu: { type: string }
                timeout: { type: integer }
                secrets: { type: array, items: { type: string } }
      responses:
        '201':
          description: Schedule created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
    get:
      summary: List schedules
      tags: [Schedules]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        '200':
          description: Schedules list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Schedule' }
  /api/schedules/{id}:
    get:
      summary: Get schedule
      tags: [Schedules]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Schedule details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
    patch:
      summary: Update schedule
      tags: [Schedules]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                schedule_value: { type: string }
                args: { type: object }
      responses:
        '200':
          description: Schedule updated
    delete:
      summary: Delete schedule
      tags: [Schedules]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Schedule deleted
  /api/schedules/{id}/trigger:
    post:
      summary: Trigger schedule manually
      tags: [Schedules]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Schedule triggered
...
  /api/apps/{id}/sandboxes:
    get:
      summary: Get app sandboxes
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sandboxes list
...